refactor: coin table 리팩토링
