+++ src/components/coinTable/const.js
@@ -1,0 +1,1 @@
+export const TABLE_HEADERS = ['동전', '개수'];
